28-oct-2002 NvE "Riostream.h" introduced to replace the standard C++ includes and                 static_cast<int> introduced in AliVertex::Draw to prevent a warning                 message for the new gcc compiler version.

+++ RALICE/AliVertex.cxx
@@ -915,7 +915,7 @@ void AliVertex::Draw(Int_t secs,Int_t cons,Int_t jets)
 
   if (!tx) continue;
 
-  charge=tx->GetCharge();
+  charge=static_cast<int>(tx->GetCharge());
 
   TPolyLine3D* line=new TPolyLine3D();
   fLines->Add(line);
